refactoring usage of props
