+++ src/components/Show/Overview/Overview.js
@@ -1,11 +1,12 @@
 import React from 'react';
 
 export const Overview = props => {
+  const { name, airDate, summary } = props;
   return (
     <div className="box">
-      <h1 className="title is-uppercase is-2">{props.name}</h1>
-      <div className="subtitle is-5 has-text-grey-light">{props.airDate.split('-')[0]}</div>
-      <div className="subtitle is-4 has-text-grey-light">{props.summary}</div>
+      <h1 className="title is-uppercase is-2">{name}</h1>
+      <div className="subtitle is-5 has-text-grey-light">{airDate.split('-')[0]}</div>
+      <div className="subtitle is-4 has-text-grey-light">{summary}</div>
     </div>
   );
 };
