+++ src/components/Show/SeasonsList/SeasonList.js
@@ -1,9 +1,10 @@
 import React from 'react';
 
 export const SeasonList = props => {
+  const { numberOfSeasons, onSelectSeason } = props;
   let seasonList = [];
 
-  for (let i = 1; i <= props.numberOfSeasons; i++) {
+  for (let i = 1; i <= numberOfSeasons; i++) {
     seasonList.push(
       <option key={i} value={i}>
         Season {i}
@@ -21,7 +22,7 @@ export const SeasonList = props => {
     <div className="field">
       <div className="control">
         <div className="select">
-          <select onChange={props.onSelectSeason}>{seasonList}</select>
+          <select onChange={onSelectSeason}>{seasonList}</select>
         </div>
       </div>
     </div>
