+++ src/components/Show/ShowCards/ShowCard/ShowCard.js
@@ -4,9 +4,10 @@ import { Link } from 'react-router-dom';
 import Image from '../../../UI/Image/Image';
 
 export const ShowCard = props => {
-  const detailLink = `/shows/${props.showId}`;
-  const image = props.imageUrl ? (
-    <Image url={props.imageUrl} alt={props.showName} />
+  const { showId, imageUrl, showName } = props;
+  const detailLink = `/shows/${showId}`;
+  const image = imageUrl ? (
+    <Image url={imageUrl} alt={showName} />
   ) : (
     <div style={{ height: '328.688px' }} />
   );
@@ -15,7 +16,7 @@ export const ShowCard = props => {
       <div className="card grow">
         <div className="card-image">{image}</div>
         <div className="card-content has-text-centered">
-          <div className="content">{props.showName}</div>
+          <div className="content">{showName}</div>
         </div>
       </div>
     </Link>
