+++ src/components/hoc/Layout/Layout.js
@@ -2,10 +2,11 @@ import React, { Fragment } from 'react';
 import { Navbar } from '../../UI/Navbar/Navbar';
 
 export const Layout = props => {
+  const { children } = props;
   return (
     <Fragment>
       <Navbar />
-      <section className="hero">{props.children}</section>
+      <section className="hero">{children}</section>
     </Fragment>
   );
 };
